# Use slices instead of cursors