--- a/src/Text/Parsing/StringParser/CodePoints.purs
+++ b/src/Text/Parsing/StringParser/CodePoints.purs
@@ -31,7 +31,7 @@ import Data.Either (Either(..))
 import Data.Enum (fromEnum)
 import Data.Foldable (class Foldable, foldMap, elem, notElem)
 import Data.Maybe (Maybe(..))
-import Data.String.CodePoints (codePointAt, drop, indexOf', length)
+import Data.String.CodePoints (codePointAt, drop, indexOf, length)
 import Data.String.CodeUnits (singleton)
 import Data.String.Pattern (Pattern(..))
 import Data.String.Regex as Regex
@@ -43,17 +43,17 @@ import Text.Parsing.StringParser.Combinators (many, (<?>))
 eof :: Parser Unit
 eof = Parser \s ->
   case s of
-    { str, pos } | pos < length str -> Left { pos, error: "Expected EOF" }
+    { substr, posFromStart } | 0 < length substr -> Left { pos: posFromStart, error: "Expected EOF" }
     _ -> Right { result: unit, suffix: s }
 
 -- | Match any character.
 anyChar :: Parser Char
-anyChar = Parser \{ str, pos } ->
-  case codePointAt pos str of
+anyChar = Parser \{ substr, posFromStart } ->
+  case codePointAt 0 substr of
     Just cp -> case toChar cp of
-      Just chr -> Right { result: chr, suffix: { str, pos: pos + 1 } }
-      Nothing -> Left { pos, error: "CodePoint " <> show cp <> " is not a character" }
-    Nothing -> Left { pos, error: "Unexpected EOF" }
+      Just chr -> Right { result: chr, suffix: { substr: drop 1 substr, posFromStart: posFromStart + 1 } }
+      Nothing -> Left { pos: posFromStart, error: "CodePoint " <> show cp <> " is not a character" }
+    Nothing -> Left { pos: posFromStart, error: "Unexpected EOF" }
   where
   toChar = fromCharCode <<< fromEnum
 
@@ -68,8 +68,8 @@ anyDigit = try do
 string :: String -> Parser String
 string nt = Parser \s ->
   case s of
-    { str, pos } | indexOf' (Pattern nt) pos str == Just pos -> Right { result: nt, suffix: { str, pos: pos + length nt } }
-    { pos } -> Left { pos, error: "Expected '" <> nt <> "'." }
+    { substr, posFromStart } | indexOf (Pattern nt) substr == Just 0 -> Right { result: nt, suffix: { substr: drop (length nt) substr, posFromStart: posFromStart + length nt } }
+    { posFromStart } -> Left { pos: posFromStart, error: "Expected '" <> nt <> "'." }
 
 -- | Match a character satisfying the given predicate.
 satisfy :: (Char -> Boolean) -> Parser Char
@@ -135,10 +135,9 @@ regex pat =
   pattern = "^(" <> pat <> ")"
 
   matchRegex :: Regex.Regex -> Parser String
-  matchRegex r = Parser \{ str, pos } -> do
-    let remainder = drop pos str
-    case NEA.head <$> Regex.match r remainder of
+  matchRegex r = Parser \{ substr, posFromStart } -> do
+    case NEA.head <$> Regex.match r substr of
       Just (Just matched) ->
-        Right { result: matched, suffix: { str, pos: pos + length matched } }
+        Right { result: matched, suffix: { substr: drop (length matched) substr, posFromStart: posFromStart + length matched } }
       _ ->
-        Left { pos, error: "no match" }
+        Left { pos: posFromStart, error: "no match" }