--- a/src/Text/Parsing/StringParser/CodePoints.purs
+++ b/src/Text/Parsing/StringParser/CodePoints.purs
@@ -43,17 +43,17 @@ import Text.Parsing.StringParser.CodeUnits as CodeUnitsParser
 eof :: Parser Unit
 eof = Parser \s ->
   case s of
-    { substr, posFromStart } | 0 < SCP.length substr -> Left { pos: posFromStart, error: "Expected EOF" }
+    { substring, position } | 0 < SCP.length substring -> Left { pos: position, error: "Expected EOF" }
     _ -> Right { result: unit, suffix: s }
 
 -- | Match any character.
 anyChar :: Parser Char
-anyChar = Parser \{ substr, posFromStart } ->
-  case SCP.codePointAt 0 substr of
+anyChar = Parser \{ substring, position } ->
+  case SCP.codePointAt 0 substring of
     Just cp -> case toChar cp of
-      Just chr -> Right { result: chr, suffix: { substr: SCP.drop 1 substr, posFromStart: posFromStart + 1 } }
-      Nothing -> Left { pos: posFromStart, error: "CodePoint " <> show cp <> " is not a character" }
-    Nothing -> Left { pos: posFromStart, error: "Unexpected EOF" }
+      Just chr -> Right { result: chr, suffix: { substring: SCP.drop 1 substring, position: position + 1 } }
+      Nothing -> Left { pos: position, error: "CodePoint " <> show cp <> " is not a character" }
+    Nothing -> Left { pos: position, error: "Unexpected EOF" }
   where
   toChar = fromCharCode <<< fromEnum
 
@@ -66,13 +66,13 @@ anyDigit = try do
 
 -- | Match the specified string.
 string :: String -> Parser String
-string pattern = Parser \{ substr, posFromStart } ->
+string pattern = Parser \{ substring, position } ->
   let
     length = SCP.length pattern
-    { before, after } = SCP.splitAt length substr
+    { before, after } = SCP.splitAt length substring
   in
-    if before == pattern then Right { result: pattern, suffix: { substr: after, posFromStart: posFromStart + length } }
-    else Left { pos: posFromStart, error: "Expected '" <> pattern <> "'." }
+    if before == pattern then Right { result: pattern, suffix: { substring: after, position: position + length } }
+    else Left { pos: position, error: "Expected '" <> pattern <> "'." }
 
 -- | Match a character satisfying the given predicate.
 satisfy :: (Char -> Boolean) -> Parser Char
@@ -138,9 +138,9 @@ regex pat =
   pattern = "^(" <> pat <> ")"
 
   matchRegex :: Regex.Regex -> Parser String
-  matchRegex r = Parser \{ substr, posFromStart } -> do
-    case NEA.head <$> Regex.match r substr of
+  matchRegex r = Parser \{ substring, position } -> do
+    case NEA.head <$> Regex.match r substring of
       Just (Just matched) ->
-        Right { result: matched, suffix: { substr: SCP.drop (SCP.length matched) substr, posFromStart: posFromStart + SCP.length matched } }
+        Right { result: matched, suffix: { substring: SCP.drop (SCP.length matched) substring, position: position + SCP.length matched } }
       _ ->
-        Left { pos: posFromStart, error: "no match" }
+        Left { pos: position, error: "no match" }