--- a/src/Text/Parsing/StringParser/Combinators.purs
+++ b/src/Text/Parsing/StringParser/Combinators.purs
@@ -21,6 +21,7 @@ module Text.Parsing.StringParser.Combinators
   , choice
   , manyTill
   , many1Till
+  , lookAhead
   , module Control.Lazy
   ) where
 
@@ -29,13 +30,21 @@ import Prelude
 import Control.Alt ((<|>))
 import Control.Lazy (fix)
 import Control.Monad.Rec.Class (Step(..), tailRecM)
+import Data.Either (Either(..))
 import Data.Foldable (class Foldable, foldl)
 import Data.List (List(..), manyRec)
 import Data.List.NonEmpty (NonEmptyList(..))
 import Data.List.NonEmpty as NEL
 import Data.Maybe (Maybe(..))
 import Data.NonEmpty ((:|))
-import Text.Parsing.StringParser (Parser, fail)
+import Text.Parsing.StringParser (Parser(..), fail)
+
+-- | Read ahead without consuming input.
+lookAhead :: forall a. Parser a -> Parser a
+lookAhead (Parser p) = Parser \s ->
+  case p s of
+    Right { result } -> Right { result, suffix: s }
+    left -> left
 
 -- | Match zero or more times.
 many :: forall a. Parser a -> Parser (List a)