… CHANGELOG.md
… test/Examples.purs
… test/Main.purs