@chtenb 2022-02-19
We probably also want to change the meaning of `posFromStart` to be `codeUnitsFromStart`. This way we can defer to only using CodeUnit operations while parsing, which will make the CodePoint and CodeUnit parsers have equivalent performance. Only when reporting the position of an error we will have to convert it to a CodePoint position.