@chtenb 2022-02-19
We can close https://github.com/purescript/purescript-strings/issues/155 once this PR gets merged