@chtenb 2022-02-27
> Why is this a breaking change, besides changing the names of the record fields in `PosString`?

Because the Parser type has changed and it is part of the public API of this package. If people have written a parser in terms of this definition, their code will have to be adjusted. Or does that not qualify as a breaking change?