@chtenb 2022-02-28
> But the only thing that has changed about the type of the Parser is the record field names in PosString, right?

Yes

I'll make an issue for anyCodePoint